Use up-to-date github PR title instead of merge-commit title
