+++ lib/getMergedPullRequests.js
@@ -1,7 +1,7 @@
 import git from 'git-promise';
 import Promise from 'bluebird';
 import semver from 'semver';
-import getPullRequestLabel from './getPullRequestLabel';
+import getGithubPullRequest from './getGithubPullRequest';
 
 function getLatestVersionTag() {
     return git('tag --list')
@@ -28,14 +28,7 @@ function getPullRequests(fromTag) {
 
 function extendWithLabel(githubRepo, validLabels, pullRequests) {
     const promises = pullRequests.map((pullRequest) => {
-        return getPullRequestLabel(githubRepo, validLabels, pullRequest.id)
-            .then((label) => {
-                return {
-                    id: pullRequest.id,
-                    title: pullRequest.title,
-                    label
-                };
-            });
+        return getGithubPullRequest(githubRepo, validLabels, pullRequest.id);
     });
 
     return Promise.all(promises);
